% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMod.R
\name{buildMod}
\alias{buildMod}
\title{buildMod}
\usage{
buildMod(
  input_data,
  vars = c("trend", "ws", "wd", "hour", "weekday", "air_temp"),
  pollutant = "NO2",
  sam.size = nrow(input_data),
  simulate = FALSE,
  B = 100,
  nrounds = nrounds,
  eta = eta,
  max_depth = max_depth,
  min_child_weight = min_child_weight,
  gamma = gamma,
  lambda = lambda,
  alpha = alpha,
  subsample = subsample,
  colsample_bytree = colsample_bytree,
  seed = seed,
  n.core = n.core,
  type = "PSOCK"
)
}
\arguments{
\item{input_data}{input_data}

\item{vars}{vars}

\item{pollutant}{pollutant}

\item{sam.size}{sam.size}

\item{simulate}{simulate}

\item{B}{B}

\item{nrounds}{nrounds}

\item{eta}{eta}

\item{max_depth}{max_depth}

\item{min_child_weight}{min_child_weight}

\item{gamma}{gamma}

\item{lambda}{lambda}

\item{alpha}{alpha}

\item{subsample}{subsample}

\item{colsample_bytree}{colsample_bytree}

\item{seed}{seed}

\item{n.core}{n.core}

\item{type}{type}
}
\description{
buildMod
}
