% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPD.R
\name{extractPD}
\alias{extractPD}
\title{extractPD}
\usage{
extractPD(vars, mod, x, n.core, type)
}
\arguments{
\item{vars}{vars}

\item{mod}{mod}

\item{x}{x}

\item{n.core}{n.core}

\item{type}{type}
}
\description{
extractPD
}
