% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPD.R
\name{getPD}
\alias{getPD}
\title{getPD}
\usage{
getPD(predictor = predictor, var = var, var_type = var_type, mod = mod, x = x)
}
\arguments{
\item{predictor}{predictor}

\item{var}{var}

\item{var_type}{var_type}

\item{mod}{mod}

\item{x}{x}
}
\description{
getPD
}
