% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metSim.R
\name{metSim}
\alias{metSim}
\title{metSim}
\usage{
metSim(
  dw_model,
  newdata,
  metVars = c("ws", "wd", "air_temp"),
  n.core = 4,
  B = 200,
  type = "multisession"
)
}
\arguments{
\item{dw_model}{mod}

\item{newdata}{data}

\item{metVars}{metVars (excl. trend)}

\item{n.core}{n.core}

\item{B}{B}

\item{type}{type multisession = SOCK, multicore = FORK}
}
\description{
metSim
}
