% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDep.R
\name{partialDep}
\alias{partialDep}
\title{partialDep}
\usage{
partialDep(
  dat,
  vars,
  B = 100,
  n.core = 4,
  nrounds = nrounds,
  eta = eta,
  max_depth = max_depth,
  min_child_weight = min_child_weight,
  gamma = gamma,
  lambda = lambda,
  alpha = alpha,
  subsample = subsample,
  colsample_bytree = colsample_bytree,
  seed = seed,
  type = "PSOCK"
)
}
\arguments{
\item{dat}{dat}

\item{vars}{vars}

\item{B}{B}

\item{n.core}{n.core}

\item{nrounds}{nrounds}

\item{eta}{eta}

\item{max_depth}{max_depth}

\item{min_child_weight}{min_child_weight}

\item{gamma}{gamma}

\item{lambda}{lambda}

\item{alpha}{alpha}

\item{subsample}{subsample}

\item{colsample_bytree}{colsample_bytree}

\item{seed}{seed}

\item{type}{type}
}
\description{
partialDep
}
