% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{prepData}
\usage{
prepData(
  mydata,
  add = c("hour", "hour.local", "weekday", "trend", "week", "jday", "month"),
  local.tz = "Europe/London",
  lag = NULL
)
}
\arguments{
\item{mydata}{data}

\item{add}{additional vars}

\item{local.tz}{local timezone}

\item{lag}{vars to lag}
}
\description{
prepData
}
