% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runXGB.R
\name{runXGB}
\alias{runXGB}
\title{RunXGB}
\usage{
runXGB(
  dat,
  vars,
  return.mod,
  simulate,
  nrounds = nrounds,
  eta = eta,
  max_depth = max_depth,
  min_child_weight = min_child_weight,
  gamma = gamma,
  lambda = lambda,
  alpha = alpha,
  subsample = subsample,
  colsample_bytree = colsample_bytree,
  seed = seed,
  n.core = n.core
)
}
\arguments{
\item{dat}{dataframe}

\item{vars}{variables}

\item{return.mod}{return model?}

\item{simulate}{random sample?}

\item{nrounds}{iterations}

\item{eta}{eta}

\item{max_depth}{max_depth}

\item{min_child_weight}{min_child_weight}

\item{gamma}{gamma}

\item{lambda}{lambda}

\item{alpha}{alpha}

\item{subsample}{subsample}

\item{colsample_bytree}{colsample_bytree}

\item{seed}{seed}

\item{n.core}{n.core}
}
\description{
RunXGB
}
